SCP-1664: Plutus Ledger API
First attempt at an interface module for the ledger team.

I had to do a lot of moving becuase we want the dependencies of this
package to be as minimal as possible, in particular we really don't want
it to depend on the plugin (which `plutus-ledger` does), or things like
`servant` or `iots-export`.